requirements.txt
paramiko
django
python-dotenv
mysql-connector-python